+++ src/grimperium/calculation/methods/standard_enthalpy_of_formation/__init__.py
@@ -1,0 +1,1 @@
+"""Standard enthalpy of formation calculation method resources."""